requirements.txt
django>=4.0.3,<4.1.0
djangorestframework>=3.13.1,<3.14.0
